feat(audio,gui,synth): add ALSA realtime backend and advanced audio tuning workflow
- add optional ALSA integration in CMake with compile-time feature gating (`ARACHNO_HAS_ALSA`)
- implement ALSA realtime output path in GUI with threaded writer queue, underrun recovery, and backend fallback
- introduce audio performance modes (Auto/Live/Balanced/Heavy/Custom) with adaptive frame-window tuning
- add Audio Performance Tuning dialog with live controls, wheel-based scaling, and custom-level presets
- calibrate default tuning profiles for higher track-count stability (including heavy-load scenarios)
- expose backend/frame telemetry in the toolbar for runtime feedback
- improve synth hot-path efficiency:
  - use float chorus buffers and optimized delay reads
  - reduce expensive math in filter cutoff path
  - rebalance heavy/extreme load thresholds
  - disable costly comb path under extreme load
- keep existing aplay path functional while allowing ALSA selection/fallback behavior

+++ include/Synthesizer.h
@@ -52,8 +52,8 @@ private:
         double lfoPhase = 0.0;
         double chorusPhase = 0.0;
         double gateSeconds = 0.25;
-        std::array<double, chorusBufferSize> chorusLeft {};
-        std::array<double, chorusBufferSize> chorusRight {};
+        std::array<float, chorusBufferSize> chorusLeft {};
+        std::array<float, chorusBufferSize> chorusRight {};
         int chorusIndex = 0;
         double crushHoldLeft = 0.0;
         double crushHoldRight = 0.0;
